Include integration tests for new search elements
